% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_factor.R
\name{wrap_factor}
\alias{wrap_factor}
\title{Ordered Factor String Wrapping}
\usage{
wrap_factor(x, n)
}
\description{
When axis text is an ordered string and needs to be wrapped while preserving the ordering
}
